requirements.txt
fastapi
uvicorn[standard]
telethon
python-dotenv
